[IMP] mail: define function for same feature of discuss actions
We'd want to standardize discuss action definition for very benefits,
such as style consistency and more UX/UI flexibility.

This commit is a small step in this direction, by defining a function
with shared commonalities of the discuss actions.
